Removed all but two Insecure Randomness findings. The last two can be mitigated.

+++ tgs/src/main/cpp/tgs/Statistics/Random.cpp
@@ -57,7 +57,7 @@ namespace Tgs
 
   double Random::generateUniform()
   {
-    return generateInt() / (double)RAND_MAX;
+    return (double)generateInt() / (double)RAND_MAX;
   }
 
   bool Random::coinToss()
